feat!: remove pdfplumber
