+++ alexi/convert.py
@@ -1,17 +1,22 @@
 """Conversion de PDF en CSV"""
 
 import csv
+import itertools
 import logging
-from collections import deque
+from os import cpu_count
 from pathlib import Path
-from typing import Any, Dict, Iterable, Iterator, Optional, TextIO
+from typing import Any, Dict, Iterable, Iterator, List, TextIO, Union
 
-from pdfplumber import PDF
-from pdfplumber.page import Page
-from pdfplumber.structure import PDFStructElement, PDFStructTree, StructTreeMissing
+import playa
+from playa.content import GlyphObject, TextObject
+from playa.pdftypes import Point, Rect
+from playa.structure import Element, Tree
+from playa.utils import get_bound_rects
 
 T_obj = Dict[str, Any]
-LOGGER = logging.getLogger("convert")
+LOGGER = logging.getLogger(__name__)
+
+
 FIELDNAMES = [
     "sequence",
     "segment",
@@ -39,120 +44,133 @@ def write_csv(
     writer.writerows(doc)
 
 
-def get_rgb(c: T_obj) -> str:
+def get_rgb(obj: TextObject) -> str:
     """Extraire la couleur d'un objet en chiffres hexadécimaux"""
-    couleur = c.get("non_stroking_color", c.get("stroking_color"))
-    if couleur is None or couleur == "":
-        return "#000"
-    elif len(couleur) == 1:
+    ncs = obj.gstate.ncs
+    ncolor = obj.gstate.ncolor
+    if ncs.ncomponents == 1:
+        gray = ncolor.values[0]
         return "#" + "".join(
-            (
-                "%x" % int(min(0.999, val) * 16)
-                for val in (couleur[0], couleur[0], couleur[0])
-            )
+            ("%x" % int(min(0.999, val) * 16) for val in (gray, gray, gray))
         )
-    elif len(couleur) == 3 or len(couleur) == 4:
-        # Could be RGB, RGBA, CMYK...
-        return "#" + "".join(("%x" % int(min(0.999, val) * 16) for val in couleur))
     else:
-        LOGGER.warning("Espace couleur non pris en charge: %s", couleur)
-        return "#000"
+        return "#" + "".join(
+            ("%x" % int(min(0.999, val) * 16) for val in ncolor.values)
+        )
 
 
-def get_word_features(
-    word: T_obj,
-    page: Page,
-    chars: dict[tuple[int, int], T_obj],
-    elmap: dict[int, str],
-) -> T_obj:
-    # Extract things from first character (we do not use
-    # extra_attrs because otherwise extract_words will
-    # insert word breaks)
-    feats = word.copy()
-    if c := chars.get((word["x0"], word["top"])):
-        feats["rgb"] = get_rgb(c)
-        mcid = c.get("mcid")
-        if mcid is not None:
-            feats["mcid"] = mcid
-            if mcid in elmap:
-                feats["tagstack"] = elmap[mcid]
-        feats["mctag"] = c.get("tag")
-        feats["fontname"] = c.get("fontname")
-    # Ensure matching PDF/CSV behaviour with missing fields
-    for field in "mcid", "sequence", "segment", "fontname", "tagstack":
-        if field not in feats or feats[field] is None:
-            feats[field] = ""
-    feats["page"] = page.page_number
-    feats["page_height"] = round(float(page.height))
-    feats["page_width"] = round(float(page.width))
-    # Round positions to points
-    for field in "x0", "x1", "top", "bottom", "doctop":
-        feats[field] = round(float(word[field]))
-    return feats
+def get_tagstack(el: Element) -> str:
+    """Extraire la généalogie d'éléments structurels."""
+    tags = [el.role]
+    parent = el.parent
+    while isinstance(parent, Element):
+        tags.append(parent.role)
+        parent = parent.parent
+    return ";".join(reversed(tags))
+
+
+def word_break(glyph: GlyphObject, origin: Point) -> bool:
+    if glyph.text == " ":
+        return True
+    x, y = glyph.origin
+    px, py = origin
+    if glyph.font.vertical:
+        off = y
+        poff = py
+    else:
+        off = x
+        poff = px
+    return off - poff > 0.5
+
+
+def line_break(glyph: GlyphObject, origin: Point) -> bool:
+    x, y = glyph.origin
+    px, py = origin
+    if glyph.font.vertical:
+        line_offset = x - px
+    else:
+        dy = y - py
+        if glyph.page.space == "screen":
+            line_offset = -dy
+        else:
+            line_offset = dy
+    return line_offset < 0 or line_offset > 100  # FIXME: arbitrary!
+
+
+def make_word(obj: TextObject, text: str, bbox: Rect) -> T_obj:
+    page = obj.page
+    x0, y0, x1, y1 = (round(x) for x in bbox)
+    word = {
+        "text": text,
+        "page": page.page_idx + 1,
+        "page_width": round(page.width),
+        "page_height": round(page.height),
+        "rgb": get_rgb(obj),
+        "x0": x0,
+        "x1": x1,
+        "top": y0,
+        "bottom": y1,
+    }
+    if obj.gstate.font is not None:
+        word["fontname"] = obj.gstate.font.fontname
+    try:
+        word["tagstack"] = get_tagstack(obj.parent)
+    except (TypeError, AttributeError):
+        pass
+    if obj.mcs is not None:
+        word["mctag"] = obj.mcs.tag
+    if obj.mcid is not None:
+        word["mcid"] = obj.mcid
+    return word
+
+
+def iter_words(objs: playa.Page) -> Iterator[T_obj]:
+    chars: List[str] = []
+    boxes: List[Rect] = []
+    textobj: Union[TextObject, None] = None
+    next_origin: Union[None, Point] = (0, 0)
+    for obj in objs:
+        if not isinstance(obj, TextObject):
+            continue
+        for glyph in obj:
+            if (
+                next_origin is not None
+                and textobj is not None
+                and chars
+                and (word_break(glyph, next_origin) or line_break(glyph, next_origin))
+            ):
+                yield make_word(textobj, "".join(chars), get_bound_rects(boxes))
+                boxes = []
+                chars = []
+                textobj = None
+            if glyph.text is not None and glyph.text != " ":
+                chars.append(glyph.text)
+                boxes.append(glyph.bbox)
+                if textobj is None:
+                    textobj = obj
+            x, y = glyph.origin
+            dx, dy = glyph.displacement
+            next_origin = (x + dx, y + dy)
+    if chars and textobj is not None:
+        yield make_word(textobj, "".join(chars), get_bound_rects(boxes))
+
+
+def extract_page(page: playa.Page) -> List[T_obj]:
+    return list(iter_words(page))
 
 
 class Converteur:
-    pdf: PDF
-    path: Path
-    tree: Optional[PDFStructTree]
-    y_tolerance: int
+    pdf: playa.Document
+    tree: Union[Tree, None]
 
-    def __init__(
-        self,
-        path: Path,
-        y_tolerance: int = 2,
-    ):
-        self.pdf = PDF.open(path)
-        self.path = path
-        self.y_tolerance = y_tolerance
-        try:
-            # Get the tree for the *entire* document since elements
-            # like the TOC may span multiple pages, and we won't find
-            # them if we look at the parent tree for other than the
-            # page in which the top element appears (this is the way
-            # the structure tree implementation in pdfplumber works,
-            # which might be a bug)
-            self.tree = PDFStructTree(self.pdf)
-        except StructTreeMissing:
-            self.tree = None
+    def __init__(self, path: Path):
+        self.pdf = playa.open(path, max_workers=round(cpu_count() / 2))
+        self.tree = self.pdf.structure
 
-    def element_map(self, page_number: int) -> dict[int, str]:
-        """Construire une correspondance entre MCIDs et types d'elements structurels"""
-        elmap: dict[int, str] = {}
-        if self.tree is None:
-            return elmap
-        d: deque[PDFStructElement | str] = deque(self.tree)
-        tagstack: deque[str] = deque()
-        while d:
-            el = d.pop()
-            if isinstance(el, str):
-                assert tagstack[-1] == el
-                tagstack.pop()
-            else:
-                d.append(el.type)
-                tagstack.append(el.type)
-                if el.page_number == page_number:
-                    for mcid in el.mcids:
-                        elmap[mcid] = ";".join(tagstack)
-                d.extend(el.children)
-        return elmap
-
-    def extract_words(self, pages: Optional[Iterable[int]] = None) -> Iterator[T_obj]:
+    def extract_words(self, pages: Union[List[int], None] = None) -> Iterator[T_obj]:
+        """Extraire mots et traits d'un PDF."""
         if pages is None:
-            pages = range(1, len(self.pdf.pages) + 1)
-        for idx in pages:
-            page = self.pdf.pages[idx - 1]
-            LOGGER.info("traitement de la page %d", page.page_number)
-            words = page.extract_words(y_tolerance=self.y_tolerance)
-            elmap = self.element_map(page.page_number)
-            # Index characters for lookup
-            chars = dict(((c["x0"], c["top"]), c) for c in page.chars)
-            for word in words:
-                if word["x0"] < 0 or word["top"] < 0:
-                    continue
-                if word["x1"] > page.width or word["bottom"] > page.height:
-                    continue
-                feats = get_word_features(word, page, chars, elmap)
-                feats["path"] = str(self.path)
-                yield feats
-            page.close()
+            pdfpages = self.pdf.pages
+        else:
+            pdfpages = self.pdf.pages[[x - 1 for x in pages]]
+        return itertools.chain.from_iterable(pdfpages.map(extract_page))
