+++ pyproject.toml
@@ -16,7 +16,6 @@ dependencies = [
     "joblib",
     "httpx",
     "lxml",
-    "pdfplumber",
     "playa-pdf>=0.6",
     "paves>=0.5",
     "scikit-learn",
