+++ scripts/retag.sh
@@ -5,7 +5,7 @@ set -e
 PDF=${1:?Usage: $0 PDF}
 CSV=${PDF/%.pdf/.csv}
 
-alexi convert --playa "$PDF" | \
+alexi convert "$PDF" | \
     alexi segment - | \
     alexi label - > "$CSV"
 open "$CSV"
