+++ scripts/view.sh
@@ -5,5 +5,5 @@ set -e
 PDF=${1:?Usage: $0 PDF}
 EXPORT=${PDF/%.pdf/.export}
 
-alexi extract -O playa -o "$EXPORT" "$PDF"
+alexi extract -o "$EXPORT" "$PDF"
 open "$EXPORT/index.html"
